requirements.txt
Django==5.2.1
python-decouple==3.8
python-dotenv==1.1.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
cloudinary==1.44.0
django-cloudinary-storage==0.3.0
psycopg==3.2.9
pandas==2.2.3
python-pptx==1.0.2
pytesseract==0.3.13
python-docx==1.1.2
sentence-transformers==4.1.0
faiss-cpu==1.11.0
accelerate==1.7.0
unstructured==0.17.2